Changes after review
1. Make PropertyMapping as marker interface with implementations: SimplePropertyMapping, CustomPropertyMapping, ReferencePropertyMapping, ReferenceMultiFieldPropertyMapping;
2. Update options in ReferenceImportPolicy: CREATE, CREATE_IF_MISSING, IGNORE_IF_MISSING, FAIL_IF_MISSING
3. Remove inheritance of ImportedDataItem from ImportedObject, put common methods to separate interface RawValuesSource.
4. Change EntityExtractor method to extract single entity: returns extracted entity only instead of entity extraction result
5. Remove property binders and replace them with different value providers: SimplePropertyValueProvider, CustomValueProvider, ReferenceValueProvider
6. Remove from ImportResult: errorType, numOfImportedEntities method
7. Replace entityMetaClassName in ImportConfiguration with entity java class

+++ dataimport/src/test/java/test_support/entity/OrderLine.java
@@ -44,7 +44,7 @@ public class OrderLine {
 
     @NotNull
     @Column(name = "QUANTITY", nullable = false, precision = 19, scale = 3)
-    private BigDecimal quantity;
+    private Integer quantity;
 
     @JoinColumn(name = "PRODUCT_ID", nullable = false)
     @NotNull
@@ -143,11 +143,11 @@ public class OrderLine {
         this.version = version;
     }
 
-    public void setQuantity(BigDecimal quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
 
-    public BigDecimal getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
